+++ src/cjs/VEllipse.js
@@ -19,6 +19,7 @@
  * @modified 2025-03-31 ATTENTION: modified the winding direction of the `tangentAt` method to match with the Circle method. This is a breaking change!
  * @modified 2025-03-31 Adding the `VEllipse.move(amount: XYCoords)` method.
  * @modified 2025-04-19 Adding the `VEllipse.getBounds()` method.
+ * @modified 2025-04-24 Adding the `VEllipse.getExtremePoints()` method for calculating minima and maxima.
  * @version  1.4.0
  *
  * @file VEllipse
@@ -118,96 +119,52 @@ var VEllipse = /** @class */ (function () {
         return new Vertex_1.Vertex(this.axis).rotate(-this.rotation, this.center).y - this.center.y;
     };
     /**
-     * Get the the northmost point of this (rotated) ellipse.
+     * Get the the minima and maxima (points) of this (rotated) ellipse.
      *
      * @method getExtremePoints
      * @instance
      * @memberof VEllipse
-     * @return {[Vertex, Vertex, Vertex, Vertex]} The "northmost" point of this (rotated) ellipse.
+     * @return {[Vertex, Vertex, Vertex, Vertex]} Get the the minima and maxima (points) of this (rotated) ellipse.
      */
     VEllipse.prototype.getExtremePoints = function () {
-        // Thanks to Cuixiping
-        //    https://stackoverflow.com/questions/87734/how-do-you-calculate-the-axis-aligned-bounding-box-of-an-ellipse
-        var r1 = this.radiusH();
-        var r2 = this.radiusV();
-        var ux = r1 * Math.cos(this.rotation);
-        var uy = r1 * Math.sin(this.rotation);
-        var vx = r2 * Math.cos(this.rotation + Math.PI / 2);
-        var vy = r2 * Math.sin(this.rotation + Math.PI / 2);
-        var c = (ux * uy + vx * vy) / (ux * vy - vx * uy);
-        var bbox_halfwidth = Math.sqrt(ux * ux + vx * vx);
-        var bbox_halfheight = Math.sqrt(uy * uy + vy * vy);
-        var x0 = c / bbox_halfwidth;
-        var y0 = c / bbox_halfheight;
-        console.log("c", c, "bbox_halfwidth", bbox_halfwidth, "x0", x0);
-        // return [new Vertex(this.center.x - ux, this.center.y - uy), new Vertex(this.center.x + vx, this.center.y + vy)];
-        var cos = Math.cos(this.rotation);
-        var sin = Math.sin(this.rotation);
-        var t1 = Math.atan((-r2 * Math.tan(this.rotation)) / r1);
-        var t2 = Math.atan((r2 * (cos / sin)) / r1);
-        // return [
-        //   new Vertex(this.center.x - bbox_halfwidth, this.center.y - bbox_halfheight),
-        //   new Vertex(this.center.x + bbox_halfwidth, this.center.y + bbox_halfheight),
-        //   new Vertex(this.center.x - bbox_halfwidth, this.center.y + bbox_halfheight),
-        //   new Vertex(this.center.x + bbox_halfwidth, this.center.y - bbox_halfheight)
-        // ];
-        return [
-            this.vertAt(-t1 - this.rotation + (0 * Math.PI) / 2.0),
-            this.vertAt(-t2 - this.rotation + (0 * Math.PI) / 2.0)
-            // new Vertex(this.center.x - x0, this.center.y - t1),
-            // new Vertex(this.center.x + x0, this.center.y + t2)
-            // new Vertex(this.center.x - bbox_halfwidth, this.center.y + uy),
-            // new Vertex(this.center.x + bbox_halfwidth, this.center.y - vy)
-        ];
-        // const ux: number = r1 * Math.cos(this.rotation);
-        // const uy: number = r1 * Math.sin(this.rotation);
-        // const vx: number = r2 * Math.cos(this.rotation + Math.PI / 2);
-        // const vy: number = r2 * Math.sin(this.rotation + Math.PI / 2);
-        // const wx: number = r2 * Math.cos(this.rotation + Math.PI);
-        // const wy: number = r2 * Math.sin(this.rotation + Math.PI);
-        // const xx: number = r2 * Math.cos(this.rotation + Math.PI * 1.5);
-        // const xy: number = r2 * Math.sin(this.rotation + Math.PI * 1.5);
-        // return [new Vertex(ux, uy).add(this.center), new Vertex(vx, vy).add(this.center)]; // , new Vertex(wx, wy), new Vertex(xx, xy)];
-        // return [
-        //   this.vertAt(-this.rotation),
-        //   this.vertAt(-this.rotation + Math.PI / 2),
-        //   this.vertAt(-this.rotation + Math.PI),
-        //   this.vertAt(-this.rotation + Math.PI * 1.5)
-        // ];
+        var a = this.radiusH();
+        var b = this.radiusV();
+        // Calculate t_x values
+        var t_x1 = Math.atan2(-b * Math.sin(this.rotation), a * Math.cos(this.rotation));
+        var t_x2 = t_x1 + Math.PI;
+        // Calculate x values at t_x
+        var x_x1 = this.center.x + a * Math.cos(t_x1) * Math.cos(this.rotation) - b * Math.sin(t_x1) * Math.sin(this.rotation);
+        var y_x1 = this.center.y + a * Math.cos(t_x1) * Math.sin(this.rotation) + b * Math.sin(t_x1) * Math.cos(this.rotation);
+        var x_x2 = this.center.x + a * Math.cos(t_x2) * Math.cos(this.rotation) - b * Math.sin(t_x2) * Math.sin(this.rotation);
+        var y_x2 = this.center.y + a * Math.cos(t_x2) * Math.sin(this.rotation) + b * Math.sin(t_x2) * Math.cos(this.rotation);
+        var x_max, x_min;
+        if (x_x1 > x_x2) {
+            x_max = new Vertex_1.Vertex(x_x1, y_x1);
+            x_min = new Vertex_1.Vertex(x_x2, y_x2);
+        }
+        else {
+            x_max = new Vertex_1.Vertex(x_x2, y_x2);
+            x_min = new Vertex_1.Vertex(x_x1, y_x1);
+        }
+        // Calculate t_y values
+        var t_y1 = Math.atan2(b * Math.cos(this.rotation), a * Math.sin(this.rotation));
+        var t_y2 = t_y1 + Math.PI;
+        // Calculate y values at t_y
+        var x_y1 = this.center.x + a * Math.cos(t_y1) * Math.cos(this.rotation) - b * Math.sin(t_y1) * Math.sin(this.rotation);
+        var y_y1 = this.center.y + a * Math.cos(t_y1) * Math.sin(this.rotation) + b * Math.sin(t_y1) * Math.cos(this.rotation);
+        var x_y2 = this.center.x + a * Math.cos(t_y2) * Math.cos(this.rotation) - b * Math.sin(t_y2) * Math.sin(this.rotation);
+        var y_y2 = this.center.y + a * Math.cos(t_y2) * Math.sin(this.rotation) + b * Math.sin(t_y2) * Math.cos(this.rotation);
+        var y_max, y_min;
+        if (y_y1 > y_y2) {
+            y_max = new Vertex_1.Vertex(x_y1, y_y1);
+            y_min = new Vertex_1.Vertex(x_y2, y_y2);
+        }
+        else {
+            y_max = new Vertex_1.Vertex(x_y2, y_y2);
+            y_min = new Vertex_1.Vertex(x_y1, y_y1);
+        }
+        return [x_max, x_min, y_max, y_min];
     };
-    // /**
-    //  * Get the center point of the south bound.
-    //  *
-    //  * @method getNorthPoint
-    //  * @instance
-    //  * @memberof Bounds
-    //  * @return {Vertex} The "southhmost" centered point of this bounding box.
-    //  */
-    // getSouthPoint(): Vertex {
-    //   return new Vertex(this.min.x + this.width / 2.0, this.max.y);
-    // }
-    // /**
-    //  * Get the center point of the west bound.
-    //  *
-    //  * @method getWestPoint
-    //  * @instance
-    //  * @memberof Bounds
-    //  * @return {Vertex} The "westhmost" centered point of this bounding box.
-    //  */
-    // getWestPoint(): Vertex {
-    //   return new Vertex(this.min.x, this.min.y + this.height / 2.0);
-    // }
-    // /**
-    //  * Get the center point of the east bound.
-    //  *
-    //  * @method getEastPoint
-    //  * @instance
-    //  * @memberof Bounds
-    //  * @return {Vertex} The "easthmost" centered point of this bounding box.
-    //  */
-    // getEastPoint(): Vertex {
-    //   return new Vertex(this.max.x, this.min.y + this.height / 2.0);
-    // }
     //--- BEGIN --- Implement interface `IBounded`
     /**
      * Get the bounds of this ellipse.
@@ -218,7 +175,7 @@ var VEllipse = /** @class */ (function () {
      * @method getBounds
      * @instance
      * @memberof VEllipse
-     * @return {Bounds} The bounds of this curve.
+     * @return {Bounds} The bounds of this ellipse.
      **/
     VEllipse.prototype.getBounds = function () {
         // Thanks to Cuixiping